UI for user typing one word at a time (#3)
Change the input so that a user types one word from the quote at
a time. Once they've typed the word correctly they advance to the
next word.  If a user types an incorrect letter, let them continue
typing and show everything after the mismatch as being in error.

New components:

- QuoteMap stores data about the quote and the position of the
  words within the quote.
- CheckedWord, so we can highlight the "current" word.
- TypingIntervals, which allow us to segment the quote into
  an already-typed interval, a mismatched interval, and an
  untyped interval.